iommu: arm-smmu: port required changes from K4.4
* Porting arm-smmu from kernel-4.4 that includes linear mapping, debug traces,
  debugfs for ptdump, regdumps, iova2phys and perf monitors.
* Updated iommu prot type to ulong to make it compatable with downstream
  implementation.
* Porting of_tegra-smmu.c from kernel-4.4 that includes device tree parsing
  required for arm-smmu driver.

Bug 1858775

Reviewed-on: http://git-master/r/1320588
GVS: Gerrit_Virtual_Submit

+++ drivers/iommu/iommu.c
@@ -1,5 +1,6 @@
 /*
  * Copyright (C) 2007-2008 Advanced Micro Devices, Inc.
+ * Copyright (c) 2017, NVIDIA CORPORATION.  All rights reserved.
  * Author: Joerg Roedel <jroedel@suse.de>
  *
  * This program is free software; you can redistribute it and/or modify it
@@ -1310,7 +1311,7 @@ static size_t iommu_pgsize(struct iommu_domain *domain,
 }
 
 int iommu_map(struct iommu_domain *domain, unsigned long iova,
-	      phys_addr_t paddr, size_t size, int prot)
+	      phys_addr_t paddr, size_t size, ulong prot)
 {
 	unsigned long orig_iova = iova;
 	unsigned int min_pagesz;
@@ -1419,7 +1420,7 @@ size_t iommu_unmap(struct iommu_domain *domain, unsigned long iova, size_t size)
 EXPORT_SYMBOL_GPL(iommu_unmap);
 
 size_t default_iommu_map_sg(struct iommu_domain *domain, unsigned long iova,
-			 struct scatterlist *sg, unsigned int nents, int prot)
+			 struct scatterlist *sg, unsigned int nents, ulong prot)
 {
 	struct scatterlist *s;
 	size_t mapped = 0;
